New Commit

+++ src/main/java/com/lti/model/UserTable.java
@@ -25,7 +25,7 @@ public class UserTable implements Serializable {
 	private String userAddress;
 
 	@Column(name="USER_AGE")
-	private BigDecimal userAge;
+	private int userAge;
 
 	@Column(name="USER_EMAIL", length=30)
 	private String userEmail;
@@ -80,11 +80,11 @@ public class UserTable implements Serializable {
 		this.userAddress = userAddress;
 	}
 
-	public BigDecimal getUserAge() {
+	public int getUserAge() {
 		return this.userAge;
 	}
 
-	public void setUserAge(BigDecimal userAge) {
+	public void setUserAge(int userAge) {
 		this.userAge = userAge;
 	}
